requirements.txt
abcplus==0.1.0
aniso8601==4.1.0
asn1crypto==0.24.0
attrs==19.1.0
bcrypt==3.1.7
cachetools==3.1.1
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
Click==7.0
colorama==0.4.1
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.7
cycler==0.10.0
Django==2.2.3
django-extra-fields==1.2.3
django-filter==2.1.0
django-rest-knox==4.1.0
django-rest-swagger==2.2.0
django-split-settings==0.3.0
djangorestframework==3.9.4
drf-yasg==1.16.1
Flask==1.0.2
Flask-RESTful==0.3.7
graphene==2.1.7
graphene-django==2.4.0
graphql-core==2.2.1
graphql-relay==2.0.0
httpie==1.0.2
idna==2.8
inflection==0.3.1
itsdangerous==1.1.0
itypes==1.1.0
Jinja2==2.10
joblib==0.13.2
jsonschema==3.0.1
kiwisolver==1.1.0
Markdown==3.1.1
MarkupSafe==1.1.0
matplotlib==3.1.0
mysqlclient==1.4.2.post1
numpy==1.15.4
openapi-codec==1.3.2
packaging==19.0
pandas==0.24.2
Pillow==6.1.0
promise==2.2.1
pycparser==2.19
Pygments==2.4.2
PyMySQL==0.9.3
pyOpenSSL==19.0.0
pyparsing==2.4.0
pyrsistent==0.15.2
python-dateutil==2.8.0
pytilhan==0.1
pytz==2018.9
PyYAML==5.1.1
requests==2.22.0
ruamel.yaml==0.15.100
Rx==1.6.1
sbm==0.3.6
schedule==0.6.0
scikit-learn==0.21.2
scipy==1.3.0
simplejson==3.16.0
singledispatch==3.4.0.3
six==1.12.0
SQLAlchemy==1.3.5
sqlparse==0.3.0
Unidecode==1.1.1
uritemplate==3.0.0
urllib3==1.25.3
virtualenv==16.6.1
Werkzeug==0.14.1
